% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_codons_to_aa.R
\name{dna_codons_to_aa}
\alias{dna_codons_to_aa}
\title{Takes a DNA codon (3 bases) and converts in into a 1-letter amino acid}
\usage{
dna_codons_to_aa(codons)
}
\description{
Takes a DNA codon (3 bases) and converts in into a 1-letter amino acid
}
