% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_rna.R
\name{random_rna}
\alias{random_rna}
\title{Generates a random rna sequence}
\usage{
random_rna(l)
}
\arguments{
\item{l}{Length of the returned rna sequence}
}
\value{
RNA sequence with length l
}
\description{
Given a length l as input, 
then generates a RNA sequence with that length
}
\examples{
# RNA with length 5
random_rna(5)

}
