% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_dna.R
\name{random_dna}
\alias{random_dna}
\title{Generates a string of random nucleotides}
\usage{
random_dna(l)
}
\arguments{
\item{l}{A number}
}
\value{
A string of random nucleotides of length \code{l}
}
\description{
Generates a string of random nucleotides
}
\examples{
random_dna(1)
random_dna()
random_dna(3)
}
